get_tracked_device_property_sys -> get_tracked_device_property
I think I can merge them into a single interface for non-MVP. Thought I
had already tried it, apparently not with the current iteration
of the trait non-MVP trait design.

It'll even be half-backwards compatible, so long as you specify return
type via `let _: type = ...` instead of via turbofish. (This means I can
write overlay code and the non-MVP variant in parallel and not have to
touch the former after the latter lands)